pyproject.toml
[tool.poetry]
name = "image_converter"
version = "0.1.0"
description = ""
authors = ["XyzzyZ4p <xyzzyzap@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.37"}
SQLAlchemy-Utils = "^0.38.2"
PyYAML = "^6.0"
asyncpg = "^0.25.0"
Pillow = "^9.1.1"
aiofiles = "^0.8.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

